Modify AccelerometerListener to send over proper timestamps
Instead of the timestamps being the time since the epoch, they are now the time
since the client(s) connected. The accuracy is either in milliseconds or
nanoseconds, depending on the API level of the device. The
CopyOnWriteArrayList<Socket> was replaced with a HashMap<Socket, Long> (note
that the container does not need to be thread-safe since it is only written from
one thread [the accepter() thread]).